Total rows now correct for map and _all_docs
We now update the Meta list with the updated total
and offset. This corrects a mistake I made when updating
to use the meta keylist, where the old total and offset
ended up getting passed to the callback.

Use of ukeysort and ukeymerge means this might get
slow if there is a lot of stuff in Meta. Might be
better to use a record?